# lib/api-zod/src/generated/types/milestoneUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Admin and bot operations for the Telegram refer-and-reward service
 * OpenAPI spec version: 0.1.0
 */

export interface MilestoneUpdate {
  name?: string;
  /** @minimum 1 */
  referralTarget?: number;
  enabled?: boolean;
}